Enable blob storage in S3
To reduce the cost and improve our reliability we want to store our
blobs in S3. Unfortunately the read/write performances are poor.

We improve the reads using a shared disk cache of the most recently used
blobs. We improve the writes by storing new blob to disk and upload
them later on using a script.

We add a BlobHelper implementation that enables to retrieve our
blobs from a S3 bucket when the blob does not exist on the shared disk.

Most of this code is coming from https://github.com/collective/collective.s3blobs

TODO:

 - tests
 - upload concurently
 - upload based on the blob content type